# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Identizen index API
  version: 0.1.0
  summary: The Identizen index — device registry, login relay, and OpenID Provider.
  description: |
    Machine-readable description of the HTTP API served by an Identizen index
    (`apps/index`). It is written by hand from the Hono routes in
    `apps/index/src/routes/*.ts`, the Zod schemas in `packages/protocol/src/schemas.ts`,
    and the error helpers in `apps/index/src/lib/errors.ts`, and it is validated by
    `apps/index/test/openapi.test.ts`: every route registered on the app must appear here and
    every path here must be served by the app. The index also serves this document as JSON at
    `GET /openapi.json`.

    Three kinds of caller use this API:

    - **Phones** (the Identizen app): register with `POST /devices`, then sign every request
      with the `Idz-Signature` header (PROTOCOL.md §8).
    - **Browsers / the SDK**: start logins (`POST /challenge`), run discovery
      (`/discover/*`), and wait on the challenge (`/challenge/{id}/state` or the WebSocket).
    - **Sites** (relying parties): standard OIDC (`/authorize`, `/token`, `/userinfo`), site
      management (`/sites`), and the server-to-server Verification API (`/v1/verify`).

    Errors are JSON `{ "error": "<code>", "error_description": "…" }`; the codes are listed
    per operation and documented at https://docs.identizen.com/errors/. Unknown routes answer
...
    headers are `Content-Type`, `Authorization`, and `Idz-Signature`).
...
servers:
...
paths:
...
  /devices:
    post:
      tags: [Devices]
      operationId: registerDevice
      summary: Register an install (and, on first sight of the master key, its identity)
      description: |
        The one unsigned phone request (PROTOCOL.md §8.1). `master_sig` is an Ed25519
        signature of type `identity` by `master_pubkey` over `{ device_pubkey, index, nonce }`,
        where `nonce` came from `POST /devices/nonce` moments earlier (valid two minutes) and `index` is this index's URL; the legacy proof over `{ device_pubkey }`
        alone is still accepted for app builds that predate nonces. A device key is enrolled at
        most once: registering an active key again returns its existing enrollment (`200`), and a
        revoked key is refused (`403 device_revoked`). The identity id `idz` is derived from
        the master public key, so a restored phone re-registers against the same identity.
        `handle` and `kind` only apply when the identity is created. Rate limited per source IP.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegistration'
      responses:
        '201':
          description: Device registered.
          content:
            application/json:
              schema:
                type: object
                required: [device_id, idz, handle, index, index_pubkey]
                properties:
                  device_id:
                    $ref: '#/components/schemas/DeviceId'
                  idz:
                    $ref: '#/components/schemas/Idz'
                  handle:
                    oneOf:
                      - $ref: '#/components/schemas/Handle'
                      - type: 'null'
                  index:
                    type: string
                    format: uri
                  index_pubkey:
                    $ref: '#/components/schemas/PublicKey'
                    description: The index's Ed25519 public key. Pin it.
        '200':
          description: This device key is already enrolled and active; its existing enrollment (same shape as `201`).
          content:
            application/json:
              schema:
                type: object
                required: [device_id, idz, handle, index, index_pubkey]
                properties:
                  device_id:
                    $ref: '#/components/schemas/DeviceId'
                  idz:
                    $ref: '#/components/schemas/Idz'
                  handle:
                    oneOf:
                      - $ref: '#/components/schemas/Handle'
                      - type: 'null'
                  index:
                    type: string
                    format: uri
                  index_pubkey:
                    $ref: '#/components/schemas/PublicKey'
        '400':
          description: '`invalid_request` — body failed validation; `bad_identity_proof` — `master_sig` does not verify over `device_pubkey` (and `index` + `nonce` when a nonce was sent); `bad_nonce` — the nonce is not one this index issued; `nonce_expired`; `invalid_destination` — a URL `push_token` the index will not call.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: '`device_revoked` — this device key was revoked or disabled; enroll with a fresh key.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: '`handle_taken` — another identity already uses the handle; `identity_mismatch` — this device key is enrolled under another identity.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /devices/nonce:
    post:
      tags: [Devices]
      operationId: registrationNonce
      summary: A nonce for the identity proof in `POST /devices`
      description: Issues a nonce, valid for two minutes, that the phone binds into its `master_sig` (PROTOCOL.md §8.1). Stateless (an HMAC under a key derived from the index signing key); replaying it is pointless because a device key is enrolled at most once. No body. Rate limited per source IP.
      responses:
        '200':
          description: The nonce and its expiry.
          content:
            application/json:
              schema:
                type: object
                required: [nonce, exp]
                properties:
                  nonce:
                    type: string
                    description: Opaque; send it back verbatim as `nonce` and inside the proof.
                  exp:
                    type: integer
                    description: Unix seconds after which the nonce is refused.
        '429':
          $ref: '#/components/responses/RateLimited'
  /devices/{id}/push-token:
    post:
      tags: [Devices]
      operationId: updatePushToken
      summary: Replace the calling device's push token
      description: Both fields are required (send `null` to clear). The signing device must be `{id}` (`403 wrong_device`).
      security:
        - idzSignature: []
      parameters:
        - $ref: '#/components/parameters/DeviceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [push_token, push_platform]
              properties:
                push_token:
                  type: [string, 'null']
                  maxLength: 4096
                  description: 'APNs / FCM token, an Expo push token (`ExponentPushToken[…]`), an HTTP(S) URL (platform `web`), or the literal `poll`.'
                push_platform:
                  oneOf:
                    - $ref: '#/components/schemas/PushPlatform'
                    - type: 'null'
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema:
                type: object
                required: [device_id, push_platform]
                properties:
                  device_id:
                    $ref: '#/components/schemas/DeviceId'
                  push_platform:
                    oneOf:
                      - $ref: '#/components/schemas/PushPlatform'
                      - type: 'null'
        '400':
          description: '`invalid_request` — body failed validation; `invalid_destination` — a URL push token the index will not call.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/DeviceUnauthorized'
        '403':
          description: '`device_inactive` — the signing device is disabled or revoked; `wrong_device` — the signature is for a different device than `{id}`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /devices/{id}/inbox:
    get:
      tags: [Devices]
      operationId: drainInbox
      summary: Drain queued challenge ids
      description: |
        The delivery of record for pushes: every challenge targeted at a device is queued here
        before any push provider is called. Returns and clears the queue (at most 50 ids;
        entries expire with the challenge). The signing device must be `{id}`.
      security:
        - idzSignature: []
      parameters:
        - $ref: '#/components/parameters/DeviceIdPath'
      responses:
        '200':
          description: Queued challenge ids, oldest first.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                type: object
                required: [challenge_ids]
                properties:
                  challenge_ids:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChallengeId'
        '401':
          $ref: '#/components/responses/DeviceUnauthorized'
        '403':
          description: '`device_inactive`; `wrong_device` — the signature is for a different device than `{id}`.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /devices/{id}/revoke:
    post:
      tags: [Devices]
      operationId: revokeDevice
      summary: Revoke a device of the same identity from another enrolled device
      description: |
        `{id}` must belong to the caller's identity. Active pairings of the device are revoked,
        its live sessions are revoked, and back-channel logout is dispatched for them. No body.
      security:
        - idzSignature: []
      parameters:
        - $ref: '#/components/parameters/DeviceIdPath'
      responses:
        '200':
          $ref: '#/components/responses/DeviceRevoked'
        '401':
          $ref: '#/components/responses/DeviceUnauthorized'
        '403':
          description: '`device_inactive` — the caller is disabled or revoked; `not_your_device` — `{id}` is unknown or belongs to another identity.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/InvalidTransition'
...
components:
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        The OIDC access token from `POST /token` (`typ: at+jwt`, ES256, 1 hour, audience =
        the index). `/userinfo` accepts any client's token. `/me/*` additionally requires the
        token to have been issued to a client listed in the index's `DASHBOARD_CLIENT_IDS`
        (`403 not_dashboard_client`) and checks the session for revocation on every call.
    idzSignature:
      type: apiKey
      in: header
      name: Idz-Signature
      description: |
        Device request authentication (PROTOCOL.md §8):

            Idz-Signature: v1,d=<device_id>,t=<unix seconds>,s=<base64url sig>

        where `sig = Ed25519.sign(deviceKey, UTF8("identizen/v1/request\n" + METHOD + "\n" +
        PATH + "\n" + base64url(SHA-256(body)) + "\n" + t))`. `METHOD` is upper-case, `PATH`
        is the request path including the query string, `body` is the raw request body (empty
        string for none). The index rejects `t` outside ±60 s of its clock (`401 bad_signature`)
        and any `(device_id, t, sig)` seen before (`401 replayed_request`); a device may send at
        most 120 signed requests per minute. `@identizen/protocol` exports `signRequest`.
    clientSecretBearer:
      type: http
      scheme: bearer
      description: |
        `Authorization: Bearer <client_secret>`. On `/sites/{client_id}` the client id comes
        from the path; on `/v1/verify*` it must be sent in the `Idz-Client-Id` header.
    clientSecretBasic:
      type: http
      scheme: basic
      description: '`Authorization: Basic base64(client_id:client_secret)`.'
    siteRegistrationToken:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <SITE_REGISTRATION_TOKEN>`; required by `POST /sites` unless the index runs with `OPEN_SITE_REGISTRATION=true`.'
  parameters:
    DeviceIdPath:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/DeviceId'
...
  responses:
...
    InvalidTransition:
      description: '`invalid_transition` — the object is already in the requested state (for example revoking a revoked device or pairing).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RateLimited:
      description: '`rate_limited` — too many requests from this IP address in a minute; `client_rate_limited` — the site started too many logins in a minute.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    DeviceUnauthorized:
      description: '`missing_signature` — `Idz-Signature` absent or malformed; `unknown_device`; `bad_signature` — wrong key, tampered request, or stale timestamp; `replayed_request`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    DeviceRevoked:
      description: The device's new status and how many live sessions were ended.
      content:
        application/json:
          schema:
            type: object
            required: [device_id, status, sessions_revoked]
            properties:
              device_id:
                $ref: '#/components/schemas/DeviceId'
              status:
                $ref: '#/components/schemas/DeviceStatus'
              sessions_revoked:
                type: integer
                minimum: 0
...
  schemas:
    Error:
      type: object
      description: Every error body. `issues` is present only for `invalid_request` from body validation.
      required: [error, error_description]
      properties:
        error:
          type: string
          description: Stable machine-readable code; see https://docs.identizen.com/errors/.
        error_description:
          type: string
        issues:
          type: array
          items:
            type: object
            required: [path, message]
            properties:
              path:
                type: string
                description: Dotted path of the offending field.
              message:
                type: string
    Base64Url:
      type: string
      pattern: '^[A-Za-z0-9_-]+$'
    PublicKey:
      type: string
      description: 32-byte Ed25519 public key, base64url (43 chars).
      pattern: '^[A-Za-z0-9_-]{43}$'
    Signature:
      type: string
      description: 64-byte Ed25519 signature, base64url (86 chars).
      pattern: '^[A-Za-z0-9_-]{86}$'
...
    KeyId:
      type: string
      description: '`base64url(SHA-256(x))[0:32]`.'
      pattern: '^[A-Za-z0-9_-]{32}$'
    Idz:
      $ref: '#/components/schemas/KeyId'
      description: Identity id — the key id of the master public key.
...
    ChallengeId:
      type: string
      pattern: '^ch_[0-9A-HJKMNP-TV-Z]{26}$'
    DeviceId:
      type: string
      pattern: '^dev_[0-9A-HJKMNP-TV-Z]{26}$'
...
    Handle:
      type: string
      minLength: 3
      maxLength: 32
      pattern: '^[a-z0-9][a-z0-9_.-]*[a-z0-9]$'
...
    IdentityKind:
      type: string
      enum: [personal, org]
    PushPlatform:
      type: string
      enum: [apns, fcm, web]
    DeviceStatus:
      type: string
      enum: [active, disabled, revoked]
...
    DeviceRegistration:
      type: object
      description: Body of `POST /devices` (unsigned).
      additionalProperties: false
      required: [device_pubkey, master_pubkey, master_sig]
      properties:
        device_pubkey:
          $ref: '#/components/schemas/PublicKey'
        master_pubkey:
          $ref: '#/components/schemas/PublicKey'
        master_sig:
          $ref: '#/components/schemas/Signature'
          description: 'Ed25519 signature of type `identity` by the master key over `{ device_pubkey, index, nonce }` when `nonce` is sent, or over `{ device_pubkey }` (legacy) when it is not.'
        nonce:
          type: string
          minLength: 40
          maxLength: 128
          description: From `POST /devices/nonce` on this index. Binds the proof to this index and this moment; omit only from app builds that predate nonces.
        handle:
          $ref: '#/components/schemas/Handle'
        kind:
          $ref: '#/components/schemas/IdentityKind'
          default: personal
        ble_key:
          $ref: '#/components/schemas/Base64Url'
          description: 32-byte BLE HMAC key, base64url. Enables BLE discovery.
        push_token:
          type: string
          maxLength: 4096
        push_platform:
          $ref: '#/components/schemas/PushPlatform'
        attestation:
          type: object
          additionalProperties: true
        label:
          type: string
          maxLength: 64
          description: Recorded in the `device.enrolled` audit event only.
...